Remove epel SLS from the generic assignment list
The epel-formula is currently coded such that it fails with an error
if run on non-Red Hat-family operating systems.  The epel SLS
assignment was previously added in a section that checked the
operating system family, but it wasn't removed from the
environment-wide list of SLS assignments.  This change corrects that
oversight.