feat: initial commit of new Apache httpd 2.4 SLS module
This is a rewrite of our pre-Git *apache* SLS module, intended to be
more generic.  We retain the ModSecurity and mod_ssl configuration but
not the OWASP ModSecurity CRS states, which has lots of special cases
not supported by the new SLS module (requiring its own SLS to be
written).  Virtual host management is also missing from this version
of the module but will be added later.