refactor: simplify package installation states
Since the `pkgs` argument to the `pkg.installed` state can be an empty
list (in which case the state completes successfully without doing
anything), we can use Jinja's `|yaml` filter to serialize each Salt
state module's package list, rather than rely on the much more
complicated combination of conditional render-time Jinja `for` loops
and state run-time requisite re-ordering.  This requires initializing
each SLS module's `packages` list.  In most cases that means setting
it to the empty list (`[]` in YAML and in Jinja), whether via
*defaults.yaml* or *map.jinja*.  At the same time, occurrences of
`require_in` and `watch_in` are replaced with `require` and `watch`
requisites in the targetted states.