fix(postfix): rework boolean type checks
Users should be able to use `yes`, `no`, `"yes"`, and `"no"` in Pillar
values, and the *postfix* formula should do the right thing.  However,
Jinja doesn't provide a way to check the type of an object.  This code
compares `value` against the boolean `True` and `False` objects before
running the other tests in an attempt to render boolean values
correctly; otherwise, `True` or `False` values print out as the strings
`"True"` or `"False"`---and results in a Postfix configuration file
syntax error.