fix(amavisd): re-order sa-update states
The `sa-update` command only needs to run once, after which time the
cron job will take care of it.  This change makes the `cmd: sa-update`
state dependent on the `cron: sa-update` state, the existence of which
then functions like a flag or a latch and prevents future invocations of
the command.