fix(apache): do not mutate in-memory data structures
Removing a key from the dictionary passed to the render_apache_config
Jinja macro ends up mutating the in-memory copy of the minion's Pillar
data in such a way that causes subsequent runs of the macro to fail to
render properly.  It is especially confusing in that if state.apply's
test kwarg equals True, the macro will appear to render correctly.  I
suspect that this is because when test=True, state functions are passed
deep copies of the relevant data, which are then discarded upon exit.

The correct solution is to filter out pre-processed keys (like
`RewriteEngine`) during directive processing, which is how the macro
handles non-directive keys (like `module_id`).