fix(auditd): disable auditd if Linux kernel support is missing
This seems to be the case with at least Raspbian 8.0 running on the
Raspberry Pi Model B, but rather than add an O/S-specific conditional,
this change implements a custom grain for Linux minions,
audit_in_linux_kernel, that will try to open the NETLINK_AUDIT socket.
If it can open the socket, in-kernel support for auditing is assumed and
the grain returns True; False, otherwise.  On non-Linux systems the
grain is not set.