feat(salt.modules.irtnog_ec2): add function that searches for the latest AMI
Because of how boto3 instantiates the boto3.ec2.Image objects, one
cannot sort a list of Image objects in Jinja by the creation_date
attribute.  This adds a function to sort the list of images in reverse
by creation date and return the most recently created AMI to the
caller.